preserve float resolution on contact_get_time and contact_post_time
server/server.py /contact GET and POST stored int(session.simulation_time)
on the Session. condition_at is float (e.g. 297.75 from the per-scenario
randomization), so an agent contacting at sim_time 297.80 was recording
contact_get_time=297, and the eval gate `contact_get_time >= condition_at`
evaluated to False (297 >= 297.75). Drop the int() cast.

server/session.py, server/schemas.py: type contact_get_time and
contact_post_time as Optional[float] to match.

eval_harness grade computes `reaction_time = stop_time - condition_at`
and formats with str() / `:.1f`, both float-safe.

+++ server/schemas.py
@@ -57,8 +57,8 @@ class EvaluateResponse(BaseModel):
     detail: str = ""
     evaluation_time: Optional[float] = None
     condition_at: Optional[float] = None
-    contact_get_time: Optional[int] = None
-    contact_post_time: Optional[int] = None
+    contact_get_time: Optional[float] = None
+    contact_post_time: Optional[float] = None
     contact_message: Optional[str] = None
